`RobotsTxtParser` does not work on Windows server
The parser uses `PHP_EOL` als a line delimiter, which on Windows is "\r\n" (two characters!).

The `newLine()` method tests for a newline on either the current char (one character so the "\r" on Windows) or on the current word. Since the current word contains all text before the newline (eg. "*\r\n") it is never triggered, resulting in an empty rules array.

This change forces the use of "\n" as a newline character in stead of PHP_EOL.